Moved dummyData to it's own file to clean up SearchPage.js, and imported dummyData into SearchPage
